+++ PINNs/paml_pinns/__init__.py
@@ -1,0 +1,2 @@
+from . import lebb as lebb
+from .train import train as train
