+++ PINNs/paml_pinns/nlebb/cases.py
