+++ PINNs/paml_pinns/nn/dense.py
