+++ PINNs/paml_pinns/nn/ffnn.py
