+++ PINNs/paml_pinns/train.py
@@ -1,15 +1,24 @@
 from functools import partial
 
 import jax
+from jaxtyping import Array
 
 import equinox as eqx
 import paramax
 import optax 
 
-from .nlebb.config import Config
 from .utils import dataloader
 
-def train(model, x, config: Config):
+
+def train(
+    model: eqx.Module,
+    x: Array,
+    weights: dict[str, float],
+    *,
+    steps: int,
+    batch_size: int = 32,
+    learning_rate: float = 1e-3
+):
 
     @partial(jax.jit, static_argnums=1)
     @jax.value_and_grad
@@ -26,14 +35,14 @@ def train(model, x, config: Config):
         return loss, params, opt_state
 
     params, static = eqx.partition(model, eqx.is_array)
-    optim = optax.adam(config.learning_rate)
+    optim = optax.adam(learning_rate)
     opt_state = optim.init(params)
 
-    iter_data = dataloader((x, x), config.batch_size) # TODO: This is ugly. Fix the data loader
+    iter_data = dataloader((x, x), batch_size) # TODO: This is ugly. Fix the data loader
 
     loss = None
-    for step, (x, _) in zip(range(config.steps), iter_data):
-        loss, params, opt_state = make_step(params, static, config.weights, x, opt_state)
+    for step, (x, _) in zip(range(steps), iter_data):
+        loss, params, opt_state = make_step(params, static, weights, x, opt_state)
         loss = loss.item()
         if step % 1000 == 0:
             print(f"Step: {step},\tLoss: {loss}")
