--- PINNs/pinns/nn/__init__.py
@@ -1,2 +1,0 @@
-from .dense import Dense
-from .ffnn import FFNN
